search of multiple words enabled on api

+++ src/main/java/com/danidemi/europrice/web/controller/api/ResourceProductItem.java
@@ -12,7 +12,7 @@ public class ResourceProductItem {
 		return this.dbProductItem.getShortDescription();
 	}
 	
-	public double getPriceInEuroCent() {
+	public long getPriceInEuroCent() {
 		return this.dbProductItem.getPriceInCent();
 	}
